requirements.txt
Django
djangorestframework
django-cors-headers
drf-extensions
